requirement.txt
adbutils==0.11.0
allure-pytest==2.9.45
allure-python-commons==2.9.45
aniso8601==9.0.1
AnyRobot==1.2.3
apipkg==1.5
apkutils2==1.0.0
appdirs==1.4.4
assume==0.4.0
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
boto3==1.16.15
botocore==1.19.15
Brotli==1.0.9
cached-property==1.5.2
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
cigam==0.0.3
click==7.1.2
colorama==0.4.4
colorlog==5.0.1
ConfigArgParse==1.5.3
configparser==5.0.2
cryptography==36.0.2
dataclasses==0.8
DBUtils==1.3
ddt==1.4.2
decorator==5.1.0
defusedxml==0.6.0
Deprecated==1.2.13
deprecation==2.1.0
distlib==0.3.1
docutils==0.16
dominate==2.6.0
execnet==1.7.1
facebook-wda==1.4.3
Faker==13.13.1
filelock==3.0.12
Flask==2.0.3
Flask-BasicAuth==0.2.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.3.3
Flask-Cors==3.0.10
flask-restplus==0.13.0
funcsigs==1.0.2
future==0.18.2
gevent==21.12.0
geventhttpclient==1.5.5
greenlet==1.1.2
idna==2.10
importlib-metadata==2.0.0
importlib-resources==5.4.0
iniconfig==1.1.1
interval==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
JPype1==1.2.1
jsonpath==0.82
jsonschema==3.2.0
lockfile==0.12.2
locust==2.8.6
logzero==1.7.0
lxml==4.6.3
MarkupSafe==2.0.1
more-itertools==8.6.0
msgpack==1.0.4
mysql==0.0.2
mysqlclient==1.4.6
numpy==1.19.5
oauthlib==3.1.0
opensearch-py==2.0.0
packaging==20.9
pandas==1.1.5
paramiko==2.11.0
Pillow==8.4.0
pluggy==0.13.1
prettytable==0.7.2
progress==1.6
psutil==5.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyecharts==1.8.1
pyelftools==0.27
Pygments==2.6.1
pymongo==3.11.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.2.4
pytest-assume==2.2.1
pytest-forked==1.1.3
pytest-ordering==0.6
pytest-repeat==0.8.0
pytest-timeout==1.4.1
pytest-xdist==1.31.0
python-daemon==2.1.2
python-dateutil==2.8.2
python-editor==1.0.4
python-slugify==4.0.1
python3-openid==3.1.0
pytz==2022.1
pywin32==302
PyYAML==5.4.1
pyzmq==23.2.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retry==0.9.2
roundrobin==0.0.2
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
selenium==3.141.0
simplejson==3.17.0
six==1.15.0
SQLAlchemy==1.3.12
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
typing==3.7.4.3
typing_extensions==4.1.1
tzlocal==1.5.1
uiautomator2==2.16.8
urllib3==1.26.4
virtualenv==20.4.4
visitor==0.1.3
wcwidth==0.2.5
weditor==0.6.4
Werkzeug==2.0.3
whichcraft==0.6.1
wrapt==1.13.2
WTForms==2.2.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0